requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
coreapi==2.3.0
coreschema==0.0.4
Django==1.9.5
django-crispy-forms==1.6.1
django-filter==1.0.2
django-formtools==2.0
django-pure-pagination==0.3.0
django-qiniu-storage==2.3.0
django-rest-swagger==2.1.2
django-reversion==2.0.8
django-simple-captcha==0.4.6
django-smtp-ssl==1.0
djangorestframework==3.6.2
DjangoUeditor==1.8.143
httplib2==0.9.2
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.1
Pillow==4.0.0
PyMySQL==0.7.10
pytz==2016.10
qiniu==7.1.1
redis==2.10.5
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
vine==1.1.3
xadmin==0.6.1
